Add support for custom backend metadata
A backend may have the need to add custom metadata to the
encoded data/parity fragments, in addition to the
liberasurecode "fragment metadata headers".  This patch
gives the ability to a EC backend writer to specify a
"metadata adder" that is taken into account when allocating
output data/parity fragment buffers.

Signed-off-by: Tushar Gohad <tushar.gohad@intel.com>