Fixing Issue #19 by stubbing out Jerasures uninit function.
If the underlying jerasure implementation is old (pre-jerasure.org),
then it will not contain an uninit function for the underlying GF
object.  Since this is only used in alg_sig, which is not used by
anything else at the moment, we stub it out if it does not exist.

Once we make the change to have alg_sig use the internal GF functions,
this whole problem goes away.