Create includedir on call to install-exec-hook
Debian maintainers reported that liberasurecode could
not be built reproducibly.

What happens is that erasurecode_version.h headers are
non-determinstically installed in the target directory
depending on the system clock. This is due to
debian/tmp/usr/include not being created and the
install-exec-hook ignores errors.

The attached patch ensures target ${includedir} exists
and therefore the headers will always be there.

Signed-off-by: Chris Lamb <lamby@debian.org>