Fix reconstruct to return an error when memory allocation failed
As like current code, if we do go jump to just "out" section when
memory allocation failed, liberaurecode/PyECLib will report it as
success. However, in fact, the returned binary is not the one
reconstructed. This patch fixes liberasurecode_reconstruct to return
error code (ENOMEM) if either malloc or memset for working space is
failed.

Change-Id: I7bfc481c7a9bbc64288760df2dc6053c57657b41