Fix liberasurecode skipping a bunch of invalid_args tests
Since the commit a01b1818c874a65d1d1fb8f11ea441e9d3e18771,
we have been to able to test some parameters for each tests.
However, instead, the NULL which means the end of the test parameters
causes skipping a bunch of unit tests which doesn't take args (e.g.
test_create_backend_invalid_args). That is the worse because we have not
tested anymore for the tests since the commit.

This patch fixes to make it tested and more, fix a minor bug for a
case that corrupted header incomming to get_metadata.

Closes-Bug: #1634403
Change-Id: Ib99a8aa6032f02d0c7d1ab94b8da1ebfd9047d74