Fix warning on automake tool
This patch fixes automake tools warnings as follows (and a my fault
in previous patch [1]):

- Change INCLUDES variable to APP_CPP_FLAGS and APP_C_FLAGS according to
  the warning message

- Remove non-POSIX variable to suppress that warnings

Note that the latter change may be affected to the testing so please
be careful on your review what's going on your testing environment.
In the past change [2], they ware designed to use the shared
object binaries in the build path. However, Daniel had pointed out the good
thing at [1] if we runs the test/liberasruecode_test (it's shell with linker)
in the dir, we can touch the expected binaries.

My fault in the previous patch is that I didn't replace
./test/.libs/liberasurecode_test to ./test/liberasurecode_test even
we use --trace-children=yes option for the valgrind.

Now, all tests run via ./test/<test name> syntax and IMO no matters
exist even if we remove the non-POSIX variable for test settings.

1: https://review.openstack.org/#/c/434162
2: 93446db9414f311bd9fc7dc047eb4dbbeb3e6feb

Change-Id: I0e79bed7755a1f286b746a70fcf56fdc972bfd5d