Allow to disable optimizations for portability
We're having trouble on Fedora when the build system runs
on Intel CPUs. The ./configure detects AVX instructions and
builds liberasurecode with them. The resulting library crashes
with SIGILL when users run it on ADM CPUs without AVX.

See the details here:
 https://bugzilla.redhat.com/show_bug.cgi?id=1454543

The patch allows to disable the optimizations, so that
distro packaging then can invoke this options and build
portable libraries.

For the record, "make test" runs about 16% slower on an Intel
CPU if optimizations are disabled. So, there's a measurable
performance impact. However, operators intersted in outright
performance might want to consider Erasure Coding implementations
other than the one  built-in into liberasurecode. The library
supports other back-ends that are significantly faster than
even optimized built-in code.

Change-Id: I09603b97ceeb833ba582cf3217e0be51c019d645