Fix write locking when creating instances
Previously, liberasurecode_backend_instance_register was the only point
where we'd do any locking. This meant that per-backend initialization
would happen outside of the lock, and we'd rely on the backend being
thread-safe. This proves a dangerous assumption; Jerasurecode is known
to not be thread-safe and seems unlikely to receive further updates.

Now, take the write lock throughout initialization.

Related-Bug: #1954351
Change-Id: I76b98fca4dce125e438566ba194e11650a2e1433