Make sure alg_sig_handle->tbl1_l is defined
On a newer GCC that runs with `-Werror=maybe-uninitialized`
liberasurecode can not be built because there isn't a guard around the
intialisation of table 1 (tbl1_l) in alg_sig_handle which causes:

  In function 'init_alg_sig_w16',
       inlined from 'init_alg_sig' at utils/chksum/alg_sig.c:242:12:
   utils/chksum/alg_sig.c:201:21: error: '*alg_sig_handle.tbl1_l' may be used uninitialized [-Werror=maybe-uninitialized]
    201 |       alg_sig_handle->tbl1_l[i] = alg_sig_handle->mult_routines.galois_single_multiply((unsigned short) (i << 8), alpha, w);
        |       ~~~~~~~~~~~~~~^~~~~~~~
    cc1: all warnings being treated as errors
    make[2]: *** [Makefile:734: utils/chksum/liberasurecode_la-alg_sig.lo]
    Error 1

This patch wraps wraps the missing if statement guard around it, so it's only
ever initialised when used.

Drive by check that initialization with malloc was actually successful.

Change-Id: Ib4c29fa56821ebd3b9472c5d379f449d31b5e104
Signed-off-by: Matthew Oliver <matt@oliver.net.au>