Move EINSUFFFRAGS checking from get_fragment_partition to caller
This allows the caller to decide the appropriate cut-off for
success, and moves the TODO about flat xor codes to
liberasurecode_reconstruct_fragment where it belongs.

Note that this *does* noticeably change the semantics of an exposed
symbol, but (to my knowledge) it was only exposed because the function
is explicitly tested. I *think* that as long as we don't mess with the
contracts of anything prefixed by "liberasurecode_" in the exposed
symbols for liberasurecode.so then we should be fine.

Change-Id: I52ed6c7fc7b37a1d9067c2d1940d203043fa24a2
Signed-off-by: Tim Burke <tim.burke@gmail.com>